chore: Change repository structure
- All paths from root directory
- motion/control.py moved to main.py
