+++ sensors/IMU.py
@@ -1,5 +1,3 @@
-#!/usr/bin/python
-
 import smbus
 import math
 import time
@@ -42,16 +40,7 @@ address = 0x68       # This is the address value read via the i2cdetect command
 # Now wake the 6050 up as it starts in sleep mode
 bus.write_byte_data(address, power_mgmt_1, 0)
 
-while True:
-    os.system('clear')
-    gyro_xout = read_word_2c(0x43)
-    gyro_yout = read_word_2c(0x45)
-    gyro_zout = read_word_2c(0x47)
-
-    print "gyro_xout : ", gyro_xout, " scaled: ", (gyro_xout / 131)
-    print "gyro_yout : ", gyro_yout, " scaled: ", (gyro_yout / 131)
-    print "gyro_zout : ", gyro_zout, " scaled: ", (gyro_zout / 131)
-
+def get_rotations():
     accel_xout = read_word_2c(0x3b)
     accel_yout = read_word_2c(0x3d)
     accel_zout = read_word_2c(0x3f)
@@ -60,11 +49,39 @@ while True:
     accel_yout_scaled = accel_yout / 16384.0
     accel_zout_scaled = accel_zout / 16384.0
 
-    print "accel_xout: ", accel_xout, " scaled: ", accel_xout_scaled
-    print "accel_yout: ", accel_yout, " scaled: ", accel_yout_scaled
-    print "accel_zout: ", accel_zout, " scaled: ", accel_zout_scaled
+    x_rot = get_x_rotation(accel_xout_scaled, accel_yout_scaled, accel_zout_scaled)
+    y_rot = get_y_rotation(accel_xout_scaled, accel_yout_scaled, accel_zout_scaled)
 
-    print "x rotation: " , get_x_rotation(accel_xout_scaled, accel_yout_scaled, accel_zout_scaled)
-    print "y rotation: " , get_y_rotation(accel_xout_scaled, accel_yout_scaled, accel_zout_scaled)
+    return (x_rot, y_rot)
 
-    time.sleep(0.5)
+def main():
+    while True:
+        os.system('clear')
+        gyro_xout = read_word_2c(0x43)
+        gyro_yout = read_word_2c(0x45)
+        gyro_zout = read_word_2c(0x47)
+
+        print("gyro_xout : ", gyro_xout, " scaled: ", (gyro_xout / 131))
+        print("gyro_yout : ", gyro_yout, " scaled: ", (gyro_yout / 131))
+        print("gyro_zout : ", gyro_zout, " scaled: ", (gyro_zout / 131))
+
+        accel_xout = read_word_2c(0x3b)
+        accel_yout = read_word_2c(0x3d)
+        accel_zout = read_word_2c(0x3f)
+
+        accel_xout_scaled = accel_xout / 16384.0
+        accel_yout_scaled = accel_yout / 16384.0
+        accel_zout_scaled = accel_zout / 16384.0
+
+        print("accel_xout: ", accel_xout, " scaled: ", accel_xout_scaled)
+        print("accel_yout: ", accel_yout, " scaled: ", accel_yout_scaled)
+        print("accel_zout: ", accel_zout, " scaled: ", accel_zout_scaled)
+
+        print("x rotation: " , get_x_rotation(accel_xout_scaled, accel_yout_scaled, accel_zout_scaled))
+        print("y rotation: " , get_y_rotation(accel_xout_scaled, accel_yout_scaled, accel_zout_scaled))
+
+
+        time.sleep(0.5)
+
+if __name__ == '__main__':
+    main()
